+++ CHANGELOG.md
@@ -44,6 +44,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `BytesMut.freeze`
 - `BytesMut.replace_with`
 - `BytesMut.offset`
+- `convert.identity`
 - New module `time` and a new class `Duration`
 - `python -m sain` command.